pyproject.toml
[tool.poetry]
name = "openzoo-server"
version = "0.1.0"
description = ""
authors = ["Ramshreyas Rao <ramshreyas.rao@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
openai = "^1.16.2"
starlette = "^0.37.2"
pydantic = "^2.6.4"
fastapi = "^0.110.1"
uvicorn = "^0.29.0"
python-dotenv = "^1.0.1"
torch = "^2.2.2"
transformers = "^4.39.3"
lamini = "^2.1.4"
sseclient-py = "^1.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
